Add schema agreement check to persistence connection
This adds a more granular level of checking for schema agreement. If the
underlying persistence implementation supports per-client schema this
allows for other custom, client-level properties to be accounted for
when checking the schema.
