+++ cql/src/main/java/org/apache/cassandra/stargate/transport/internal/messages/QueryMessage.java
@@ -72,7 +72,7 @@ public class QueryMessage extends Message.Request {
     SimpleStatement statement = new SimpleStatement(query, options.getValues(), options.getNames());
     CompletableFuture<? extends Result> future =
         persistenceConnection().execute(statement, makeParameters(options), queryStartNanoTime);
-    return SchemaAgreement.maybeWaitForAgreement(future, persistence())
+    return SchemaAgreement.maybeWaitForAgreement(future, persistenceConnection())
         .thenApply(ResultMessage::new);
   }
