+++ persistence-api/src/main/java/io/stargate/db/Persistence.java
@@ -15,7 +15,6 @@
  */
 package io.stargate.db;
 
-import com.datastax.oss.driver.shaded.guava.common.util.concurrent.Uninterruptibles;
 import io.stargate.db.datastore.DataStore;
 import io.stargate.db.schema.Schema;
 import io.stargate.db.schema.TableName;
@@ -24,7 +23,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
 import org.apache.cassandra.stargate.exceptions.AuthenticationException;
 
 /**
@@ -34,10 +32,7 @@ import org.apache.cassandra.stargate.exceptions.AuthenticationException;
  * higher level {@link DataStore} API which wraps an instance of this interface) to query the
  * underlying store, and thus the one interface that persistence extensions must implement.
  */
-public interface Persistence {
-
-  int SCHEMA_AGREEMENT_WAIT_RETRIES =
-      Integer.getInteger("stargate.persistence.schema.agreement.wait.retries", 900);
+public interface Persistence extends SchemaChecker {
 
   /** Name describing the persistence implementation. */
   String name();
@@ -105,20 +100,6 @@ public interface Persistence {
     return true;
   }
 
-  /** Wait for schema to agree across the cluster */
-  default void waitForSchemaAgreement() {
-    for (int count = 0; count < SCHEMA_AGREEMENT_WAIT_RETRIES; count++) {
-      if (isInSchemaAgreement()) {
-        return;
-      }
-      Uninterruptibles.sleepUninterruptibly(200, TimeUnit.MILLISECONDS);
-    }
-    throw new IllegalStateException(
-        "Failed to reach schema agreement after "
-            + (200 * SCHEMA_AGREEMENT_WAIT_RETRIES)
-            + " milliseconds.");
-  }
-
   /**
    * Returns <code>false</code>> if the persistence implementation cannot reasonably expect to reach
    * schema agreement with storage nodes assuming no external intervention.
@@ -155,7 +136,7 @@ public interface Persistence {
    * <p>It is through this object that a user can be logged in and that the persistence can be
    * queried.
    */
-  interface Connection {
+  interface Connection extends SchemaChecker {
 
     /** The underlying persistence on which this is a connection. */
     Persistence persistence();
@@ -252,5 +233,16 @@ public interface Persistence {
 
     /** Makes a new {@link RowDecorator} for the given table. */
     RowDecorator makeRowDecorator(TableName table);
+
+    /**
+     * Checks whether this coordinator in schema agreement with the other nodes in the cluster. This
+     * is the same as {@code Persistence#isInSchemaAgreement()}, except that it *may* consider other
+     * connection specific properties to determine whether schema is in agreement.
+     *
+     * @return true if schema is agreement; otherwise false.
+     */
+    default boolean isInSchemaAgreement() {
+      return persistence().isInSchemaAgreement();
+    }
   }
 }
