+++ persistence-api/src/main/java/io/stargate/db/datastore/PersistenceBackedDataStore.java
@@ -221,7 +221,7 @@ public class PersistenceBackedDataStore implements DataStore {
 
   @Override
   public void waitForSchemaAgreement() {
-    persistence().waitForSchemaAgreement();
+    connection.waitForSchemaAgreement();
   }
 
   @Override
